+++ CHANGES.rst
@@ -6,6 +6,7 @@ IP Navigator CHANGES
 Development
 ===========
 - [ui] Upgrade to jQuery 3.4.1
+- [ui] Add "clean-webpack-plugin" to cleanup the assets folder before building
 
 
 2019-05-08 0.165.0